requirements.txt
kafka-python==2.0.2
python-dotenv==1.0.1
confluent_kafka==2.5.3